IMPALA-13194: Fast-serialize position delete records
Currently the serialization of position delete records are very
wasteful. The records contain slots 'file_path' and 'pos'. And
what we do during serialization is the following:

 1. Write fixed-size tuples that have a StringValue and a BigInt slot
 2. Copy the StringValue's contents after the tuple.
 3. Convert the StringValue ptr to be an offset to the string data

So we end up having something like this:

+-------------+--------+----------------+-------------+--------+----------------+-----+
| StringValue | BigInt |   File path    | StringValue | BigInt |   File path    | ... |
+-------------+--------+----------------+-------------+--------+----------------+-----+
| ptr, len    |     42 | /.../a.parquet | ptr, len    |     43 | /.../a.parquet | ... |
+-------------+--------+----------------+-------------+--------+----------------+-----+

This is very redundant to store the file paths that way, and in the
end we will have a huge buffer that we need to compress and send over
the network. Moreover, we copy the file paths in memory twice:

 1. From input row batch to the KrpcDataStreamSender::Channel's temporary row batch
 2. From the temporary row batch to the outbound row batch (during serialization)

The position delete files store the delete records in ascending order.
This means adjacent records mostly have the same file path. So we could
just buffer the position delete records up to the Channel's capacity,
then serialize the data in a more efficient way.

With this patch, serialized data will look like this:
+----------------+-------------+--------+-------------+--------+-----+
|   File path    | StringValue | BigInt | StringValue | BigInt | ... |
+----------------+-------------+--------+-------------+--------+-----+
| /.../a.parquet | ptr, len    |     42 | ptr, len    |     43 | ... |
+----------------+-------------+--------+-------------+--------+-----+

File path, then tuples with the same file path, after that comes the
next file path and tuples associated with that one, and so on.

Measurements:
07:EXCHANGE        : 1m   ==> 52s
F02:EXCHANGE SENDER: 1m2s ==> 16s

Change-Id: I6095f318e3d06dedb4197681156b40dd2a326c6f
Reviewed-on: http://gerrit.cloudera.org:8080/21563
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>