IMPALA-13189: Reset the database on clearing imported query profiles
On clicking the "X" button beside the imported query profiles
section header, deleting the entire "imported_profiles" database
and recreating it instead of clearing the "profiles" object store
would be more beneficial.

This could potentially resolve problems with indexing in the database.

When the "imported_profiles" database is being deleted, a message is
displayed adjacent to the imported query profiles section header
to ensure page is not reloaded / closed during the process.

Change-Id: I0c70b4aaf66f6dc3804a279928d610475834e4f3
Reviewed-on: http://gerrit.cloudera.org:8080/21558
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>