IMPALA-13197: Implement Analytic Exprs for Calcite
An analytic expression is represented with a RexOver type RexNode
within Calcite. They will exist within the Project RelNode. If
there are any RexOvers existing within the Project, then the
ImpalaAnalyticRel RelNode gets created instead of the ImpalaProjectRel.

Only bare bones test cases are included. There are quite a number
of analytic expressions that will not work. The logic is included in
the AnalyticExpr.standardize() method. Another commit will be needed
to support all general analytic expressions and the tests within
Impala will be used for testing purposes.

Change-Id: Iba5060546a7568ba0cd315f546daa78d89b1c3c5
Reviewed-on: http://gerrit.cloudera.org:8080/21565
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>