IMPALA-13226: Rename TupleCacheInfo.finalize() to finalizeHash()
TupleCacheInfo.finalize() unintentionally overwrites Object.finalize()
which is called by the JVM garbage collector when garbage collection
determines that there are no more references to the object. Usually the
finalize method is overrided to dispose of system resources or to
perform other cleanup.

TupleCacheInfo.finalize() is not meant to be used during GC. We'd better
use another method name to avoid confusion. This patch renames it to
finalizeHash(). Also fixed some stale comments.

Change-Id: I657c4f14b074b7c16dc7d126b0c8b5083b8f19c6
Reviewed-on: http://gerrit.cloudera.org:8080/21588
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>