IMPALA-13252: Always use PrintId for TUniqueId
Thrift types provide their own overload of operator<<. In GCC, the
overload added in debug-util.h is preferred, but in Clang the first
overload (from Thrift generated code) is preferred. I don't see a good
way to augment the Thrift definition to use PrintId, so this patch
reverts the overload in debug-util.h and uses PrintId explicitly.

Change-Id: Ic536c12ed67d9d473c7ac7f1e05ebc537ff485d4
Reviewed-on: http://gerrit.cloudera.org:8080/21639
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Yida Wu <wydbaggio000@gmail.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>