IMPALA-13272: Analytic function of collections can lead to crash
The following query leads to DCHECK in debug builds and may cause more
subtle issues in RELEASE builds:

  select
    row_no
  from (
    select
      arr.small,
      row_number() over (order by arr.inner_struct1.str) as row_no
    from functional_parquet.collection_struct_mix t,
        t.arr_contains_nested_struct arr
  ) res;

The problem is in AnalyticPlanner.createSortInfo(). Because it is an
array unnesting operation, there are two tuples from which we try to add
slot descriptors to the sorting tuple: the array item tuple (which we'll
need) and the main tuple (which we don't actually need). The main tuple
contains the slot desc for the array. It is marked as materialised, so
we add it to the sorting tuple, but its child 'small' is not
materialised, which leads to the error.

This change solves the problem by only adding slot descs to the sorting
tuple if they are fully materialised, i.e. they and all their children
recursively are also materialised.

Testing:
 - added test queries in sort-complex.test.

Change-Id: I71d1fa28ad4ff2e1a8fc5b91d3fc271c33765190
Reviewed-on: http://gerrit.cloudera.org:8080/21643
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>