IMPALA-12737: Add HS2 support to the InternalServer class.
The existing methods in the InternalServer class allow for retrieving
query results via tab delimited text. This method is not robust when
accuracy is needed for query results. This commit adds the ability to
use the HS2 protocol to execute queries with the results returned as
a vector of Thrift TRow objects. This new method is not used in this
commit but will be used in subsequent commits for IMPALA-12737.

Testing was accomplished by adding a new unit ctest.

Change-Id: Idf4b278b69ffb2beef3ea5bc1eb9335d9c27a5c8
Reviewed-on: http://gerrit.cloudera.org:8080/21650
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>