IMPALA-12758: Fix catalogd not setting prev_id for reloaded partitions
Cache modification via partitioned table events should retain the
prev_id of the partitions while reloading the partitions. Currently,
prev_id is set to -1, so when the catalogDelta is sent from the
statestore to impala demons, because prev_id is not valid, coordinators
in local catalog mode will not know whether to invalidate the current
partition.

This patch addresses this issue by setting the prev_id of reloaded
partition.

Note: This is only an issue in local catalog mode for
REFRESH <partition> or updating partitions based on HMS events.

Testing:
- Added an end-to-end test to verify the same

Change-Id: Ia7d53f601b63d83e99cbd2d1f58076a0dc78c610
Reviewed-on: http://gerrit.cloudera.org:8080/21662
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>