IMPALA-12865: Fix wrong lastRefreshEventId set by firing RELOAD events
When enable_reload_events is true, catalogd fires RELOAD events after
REFRESH finishes reloading the table/partition. The RELOAD event id is
also used to update lastRefreshEventId of the table/partition. This is
problematic when enable_skipping_older_events is true. HMS events
generated after the reload and before the RELOAD event will be skipped.

Solution:
Fetch the current HMS notification event id before the table/partition
is refreshed, and set it as lastRefreshEventId on the metadata object.

Testing:
- Manually verified the issue is addressed.
- Added an end-to-end test which is close to real time issue.

Change-Id: I90039da77ec561c5aede44456f88c6650582815b
Reviewed-on: http://gerrit.cloudera.org:8080/21665
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>