IMPALA-13300: Upgrade Maven to 3.9.8
Maven version 3.9.7 consumed an upgraded version of the resolver
plugin that contains a fix around file locking. Issues with locking
files are seen occasionally on builds.

This patch consumes Maven 3.9.8 since it is the latest version
available at this time.

Testing was performed by running only the download code in a Redhat 8
docker container.

Change-Id: I509dd94799b99bf637a583eadc2905bc32a87c87
Reviewed-on: http://gerrit.cloudera.org:8080/21674
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Andrew Sherman <asherman@cloudera.com>