IMPALA-13296: Check column compatibility earlier for table migration
When migrating a Hive table to Iceberg there can be columns in the
source table that are not supported by Iceberg. For these tables the
table migration process gives an error. However, this is a multi-step
process and the check for column compatibility is performed after the
source table had been renamed to a temporary name.
This patch brings the column type check to the analysis phase to avoid
renaming the table when there are incompatible columns.

Testing:
  - Re-run the existing checks
  - Adjusted the error message for this error case to reflect that we
    return an AnalysisException.

Change-Id: I381e7359a55ebd84f9a1dcec3b665bdbea62a035
Reviewed-on: http://gerrit.cloudera.org:8080/21675
Reviewed-by: Peter Rozsa <prozsa@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>