IMPALA-13298: Correct test_miss_complete_cb Assertion
Commit ad868b9 modified the RPC error messages to include the
query id. Part of this change was to remove the string
"Query aborted" from the test's beeswax client when it raises an
error.

The TestRPCTimeout.test_miss_complete_cb was missed and was still
asserting on the string "Query aborted". This patch fixes that test
to use the new error_msg_expected() function.

Testing was accomplished by running the
TestRPCTimeout.test_miss_complete_cb test locally.

Change-Id: Ifb2fa3864b52b0927c6488b22135913b90bce3bd
Reviewed-on: http://gerrit.cloudera.org:8080/21678
Reviewed-by: Riza Suminto <riza.suminto@cloudera.com>
Reviewed-by: Xuebin Su <xsu@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Michael Smith <michael.smith@cloudera.com>