IMPALA-13303: FileSystemUtil.listFiles() should handle non-recursive case
FileSystemUtil.listFiles() is used in FileMetadataLoader#loadInternal()
to list the files with block locations. When table property
"impala.disable.recursive.listing" is set to true, it's supposed to skip
files in the sub dirs. However, for FileSystems that don't support
recursive listFiles(), we always create a RecursingIterator and don't
respect the 'recursive' argument.

This patch fixes the issue by adding the check for the 'recursive'
argument and use the non-recursive iterator when it's false.

Tests
 - Add test in test_recursive_listing.py to reveal the issue

Change-Id: Ia930e6071963d53561ce79896bff9d19720468a4
Reviewed-on: http://gerrit.cloudera.org:8080/21680
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>