IMPALA-12363: Upgrade RE2 to 2023-03-01
This bumps the version of re2 to 2023-03-01, which is the
last release that doesn't have an Abseil dependency.
The toolchain already contains a build of re2 2023-03-01,
so there is no need to bump the toolchain version.

This has a performance benefit for TPC-H's Q13, which uses
this predicate: "o_comment not like '%special%requests%"
This like predicate is complicated enough that it doesn't fit
the heavily optimized paths that exist for simpler likes.
Instead, this gets converted to an RE2 regex. The newer RE2
significantly improves performance of that predicate, and
TPC-H Q13 gets ~9% faster.

Testing:
 - Ran a core job
 - Ran a perf-AB-test

Change-Id: Ic7f131102bd7590d222f22dcc412d9fd2286f006
Reviewed-on: http://gerrit.cloudera.org:8080/21712
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Yida Wu <wydbaggio000@gmail.com>