IMPALA-13082: Use separate versions for jackson vs jackson-databind
Sometimes there is a jackson-databind patch release without
a corresponding release of other jackson libraries. For example,
there is a jackson-databind 2.12.7.1, but jackson-core does not
have an artifact with that version. To handle these scenarios,
it is useful to have a separate version for jackson-databind
vs other jackson libraries.

This introduces IMPALA_JACKSON_VERSION (which currently matches
IMPALA_JACKSON_DATABIND_VERSION) and uses this for non-databind
jackson libraries.

Testing:
 - Ran a local build

Change-Id: I3055cb47986581793d947eaedb6a24b4dd92e3a6
Reviewed-on: http://gerrit.cloudera.org:8080/21719
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>