IMPALA-13336: Fix syntax error in creating Iceberg test table on Apache Hive 3
Apache Hive 3 doesn't support the syntax of STORED BY ICEBERG STORED AS
AVRO. When loading test data on Apache Hive 3, we convert this clause to
  STORED BY 'org.apache.iceberg.mr.hive.HiveIcebergStorageHandler'
  TBLPROPERTIES('write.format.default'='AVRO')
However, when there is a LOCATION clause in the statement, the
TBLPROPERTIES clause will be put before the LOCATION clause, which
causes the syntax error.

In the CreateTable statement, TBLPROPERTIES clause should be put after
the LOCATION clause. This patch fixes generate-schema-statements.py to
take care of this case.

Tests:
 - Verified the SQL files generated by generate-schema-statements.py

Change-Id: I5b47d6dc1a2ab63d4ecea476dbab67c1ae8ca490
Reviewed-on: http://gerrit.cloudera.org:8080/21730
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>