IMPALA-11431: Avoid getting stats for synthetic column row__id from HMS
Before this change Impala always tried to fetch stats for row__id in
full ACID tables, which does not exist in the metastore. Sometimes this
led to an exception from HMS, see HIVE-28498 for details.

This caused flakyness in test_compute_stats_with_structs. The test also
had side effects (it computed stats for a shared table) so it was
modified to use unique_database.

Testing:
- could reproduce the issue by starting HMS with
  hive.metastore.try.direct.sql=true and verified that the change
  fixes it

Change-Id: I759f57c99aa16e4ab5fd82aa5f6b756446291f03
Reviewed-on: http://gerrit.cloudera.org:8080/21742
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>